backend/app/auth/jwt.py
from typing import Any
from fastapi import HTTPException, Header
from jose import jwt, jwk, JWTError
from app.config import settings
import httpx

_jwks_cache: list | None = None


def _get_jwks() -> Any:
    global _jwks_cache
    if _jwks_cache is None:
        res = httpx.get(f"{settings.supabase_url}/auth/v1/.well-known/jwks.json")
        res.raise_for_status()
        _jwks_cache = res.json()["keys"]
    return _jwks_cache


def get_current_user(authorization: str = Header(...)) -> str:
    if not authorization.startswith("Bearer "):
        raise HTTPException(
            status_code=401, detail="Invalid authorization header format"
        )

    token = authorization.removeprefix("Bearer ")

    try:
        keys = _get_jwks()
        payload = None
        for key_data in keys:
            try:
                public_key = jwk.construct(key_data)
                payload = jwt.decode(
                    token, public_key, algorithms=["ES256"], audience="authenticated"
                )
                break
            except JWTError:
                continue

        if payload is None:
            raise JWTError("No valid key found")

    except JWTError as e:
        raise HTTPException(status_code=401, detail=f"Invalid token: {str(e)}")

    user_id: str = str(payload.get("sub"))
    if not user_id:
        raise HTTPException(status_code=401, detail="Token missing user ID")

    return user_id
