backend/pyproject.toml
[project]
name = "sqlo-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.135,<1.0",
  "uvicorn>=0.35,<1.0",
  "supabase>=2.0,<3.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "pydantic>=2.11,<3.0",
  "pydantic-settings>=2.0,<3.0",
  "python-dotenv>=1.0,<2.0",
  "openai>=1.0,<2.0",
  "faker>=25.0,<26.0",
  "numpy>=2.0,<3.0",
  "httpx>=0.27,<1.0",
  "anthropic>=0.90.0",
]

[project.optional-dependencies]
dev = ["pytest"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = ["pytest>=9.0.3"]
